refactor group views
